Isolate devnet config, move parsing, dump synchronously
