+++ starknet_devnet/server.py
@@ -16,9 +16,9 @@ from .blueprints.gateway import gateway
 from .blueprints.feeder_gateway import feeder_gateway
 from .blueprints.postman import postman
 from .blueprints.rpc import rpc
-from .util import DumpOn, check_valid_dump_path, parse_args
-from .starknet_wrapper import DevnetConfig
+from .util import check_valid_dump_path
 from .state import state
+from .devnet_config import devnet_config, DumpOn
 
 app = Flask(__name__)
 CORS(app)
@@ -33,16 +33,6 @@ app.register_blueprint(gateway)
 app.register_blueprint(feeder_gateway)
 app.register_blueprint(postman)
 app.register_blueprint(rpc)
-
-def generate_accounts(args):
-    """Generate accounts """
-    if args.accounts:
-        state.starknet_wrapper.accounts.generate(
-            n_accounts=args.accounts,
-            initial_balance=args.initial_balance,
-            seed=args.seed
-        )
-        state.starknet_wrapper.accounts.print()
 
 def set_dump_options(args):
     """Assign dumping options from args to state."""
@@ -62,30 +52,6 @@ def load_dumped(args):
             state.load(args.load_path)
         except (FileNotFoundError, UnpicklingError):
             sys.exit(f"Error: Cannot load from {args.load_path}. Make sure the file exists and contains a Devnet dump.")
-
-def enable_lite_mode(args):
-    """Enable lite mode if specified."""
-    if args.lite_mode:
-        config = DevnetConfig(
-            lite_mode_block_hash=True,
-            lite_mode_deploy_hash=True
-        )
-    else:
-        config = DevnetConfig(
-            lite_mode_block_hash=args.lite_mode_block_hash,
-            lite_mode_deploy_hash=args.lite_mode_deploy_hash
-        )
-
-    state.starknet_wrapper.set_config(config)
-
-def set_start_time(args):
-    """Assign start time if specified."""
-    if args.start_time is not None:
-        state.starknet_wrapper.set_block_time(args.start_time)
-
-def set_gas_price(args):
-    """Assign gas_price"""
-    state.starknet_wrapper.set_gas_price(args.gas_price)
 
 # We don't need init method here.
 # pylint: disable=W0223
@@ -127,18 +93,13 @@ class Devnet(BaseApplication):
 def main():
     """Runs the server."""
 
-    args = parse_args()
-
     # Uncomment this once fork support is added
     # origin = Origin(args.fork) if args.fork else NullOrigin()
     # starknet_wrapper.origin = origin
 
+    args = devnet_config.args
     load_dumped(args)
     set_dump_options(args)
-    generate_accounts(args)
-    enable_lite_mode(args)
-    set_start_time(args)
-    set_gas_price(args)
 
     asyncio.run(state.starknet_wrapper.initialize())
